Updating foxhound version

+++ ipc/chromium/src/base/process_util_posix.cc
@@ -30,6 +30,7 @@
 #include "base/dir_reader_posix.h"
 
 #include "mozilla/UniquePtr.h"
+#include "mozilla/Unused.h"
 // For PR_DuplicateEnvironment:
 #include "prenv.h"
 #include "prmem.h"
@@ -38,7 +39,13 @@
 #  include "mozilla/ipc/ForkServiceChild.h"
 #endif
 
-const int kMicrosecondsPerSecond = 1000000;
+// We could configure-test for `waitid`, but it's been in POSIX for a
+// long time and OpenBSD seems to be the only Unix we target that
+// doesn't have it.  Note that `waitid` is used to resolve a conflict
+// with the crash reporter, which isn't available on OpenBSD.
+#ifndef __OpenBSD__
+#  define HAVE_WAITID
+#endif
 
 namespace base {
 
@@ -114,13 +121,14 @@ void CloseSuperfluousFds(void* aCtx, bool (*aShouldPreserve)(void*, int)) {
 #if defined(ANDROID)
   static const rlim_t kSystemDefaultMaxFds = 1024;
   static const char kFDDir[] = "/proc/self/fd";
-#elif defined(OS_LINUX) || defined(OS_SOLARIS)
+#elif defined(XP_LINUX) || defined(XP_SOLARIS)
   static const rlim_t kSystemDefaultMaxFds = 8192;
   static const char kFDDir[] = "/proc/self/fd";
-#elif defined(OS_MACOSX)
+#elif defined(XP_DARWIN)
   static const rlim_t kSystemDefaultMaxFds = 256;
   static const char kFDDir[] = "/dev/fd";
-#elif defined(OS_BSD)
+#elif defined(__DragonFly__) || defined(XP_FREEBSD) || defined(XP_NETBSD) || \
+    defined(XP_OPENBSD)
   // the getrlimit below should never fail, so whatever ..
   static const rlim_t kSystemDefaultMaxFds = 1024;
   // at least /dev/fd will exist
@@ -187,27 +195,42 @@ void CloseSuperfluousFds(void* aCtx, bool (*aShouldPreserve)(void*, int)) {
   }
 }
 
-bool DidProcessCrash(bool* child_exited, ProcessHandle handle) {
+bool IsProcessDead(ProcessHandle handle, bool blocking) {
+  auto handleForkServer = [handle]() -> mozilla::Maybe<bool> {
 #ifdef MOZ_ENABLE_FORKSERVER
-  if (mozilla::ipc::ForkServiceChild::Get()) {
-    // We only know if a process exists, but not if it has crashed.
-    //
-    // Since content processes are not direct children of the chrome
-    // process any more, it is impossible to use |waitpid()| to wait for
-    // them.
-    const int r = kill(handle, 0);
-    if (r < 0 && errno == ESRCH) {
-      if (child_exited) *child_exited = true;
-    } else {
-      if (child_exited) *child_exited = false;
+    if (errno == ECHILD && mozilla::ipc::ForkServiceChild::Get()) {
+      // We only know if a process exists, but not if it has crashed.
+      //
+      // Since content processes are not direct children of the chrome
+      // process any more, it is impossible to use |waitpid()| to wait for
+      // them.
+      const int r = kill(handle, 0);
+      // FIXME: for unexpected errors we should probably log a warning
+      // and return true, so that the caller doesn't loop / hang /
+      // try to kill the process.  (Bug 1658072 will rewrite this code.)
+      return mozilla::Some(r < 0 && errno == ESRCH);
+    }
+#else
+    mozilla::Unused << handle;
+#endif
+    return mozilla::Nothing();
+  };
+
+#ifdef HAVE_WAITID
+
+  // We use `WNOWAIT` to read the process status without
+  // side-effecting it, in case it's something unexpected like a
+  // ptrace-stop for the crash reporter.  If is an exit, the call is
+  // reissued (see the end of this function) without that flag in
+  // order to collect the process.
+  siginfo_t si{};
+  const int wflags = WEXITED | WNOWAIT | (blocking ? 0 : WNOHANG);
+  int result = HANDLE_EINTR(waitid(P_PID, handle, &si, wflags));
+  if (result == -1) {
+    if (auto forkServerReturn = handleForkServer()) {
+      return *forkServerReturn;
     }
 
-    return false;
-  }
-#endif
-  int status;
-  const int result = HANDLE_EINTR(waitpid(handle, &status, WNOHANG));
-  if (result == -1) {
     // This shouldn't happen, but sometimes it does.  The error is
     // probably ECHILD and the reason is probably that a pid was
     // waited on again after a previous wait reclaimed its zombie.
@@ -218,34 +241,83 @@ bool DidProcessCrash(bool* child_exited, ProcessHandle handle) {
     // So, lacking reliable information, we indicate that the process
     // is dead, in the hope that the caller will give up and stop
     // calling us.  See also bug 943174 and bug 933680.
+    CHROMIUM_LOG(ERROR) << "waitid failed pid:" << handle << " errno:" << errno;
+    return true;
+  }
+
+  if (si.si_pid == 0) {
+    // the child hasn't exited yet.
+    return false;
+  }
+
+  DCHECK(si.si_pid == handle);
+  switch (si.si_code) {
+    case CLD_STOPPED:
+    case CLD_CONTINUED:
+      DCHECK(false) << "waitid returned an event type that it shouldn't have";
+      [[fallthrough]];
+    case CLD_TRAPPED:
+      CHROMIUM_LOG(WARNING) << "ignoring non-exit event for process " << handle;
+      return false;
+
+    case CLD_KILLED:
+    case CLD_DUMPED:
+      CHROMIUM_LOG(WARNING)
+          << "process " << handle << " exited on signal " << si.si_status;
+      break;
+
+    case CLD_EXITED:
+      if (si.si_status != 0) {
+        CHROMIUM_LOG(WARNING)
+            << "process " << handle << " exited with status " << si.si_status;
+      }
+      break;
+
+    default:
+      CHROMIUM_LOG(ERROR) << "unexpected waitid si_code value: " << si.si_code;
+      DCHECK(false);
+      // This shouldn't happen, but assume that the process exited to
+      // avoid the caller possibly ending up in a loop.
+  }
+
+  // Now consume the status / collect the dead process
+  const int old_si_code = si.si_code;
+  si.si_pid = 0;
+  // In theory it shouldn't matter either way if we use `WNOHANG` at
+  // this point, but just in case, avoid unexpected blocking.
+  result = HANDLE_EINTR(waitid(P_PID, handle, &si, WEXITED | WNOHANG));
+  DCHECK(result == 0);
+  DCHECK(si.si_pid == handle);
+  DCHECK(si.si_code == old_si_code);
+  return true;
+
+#else  // no waitid
+
+  int status;
+  const int result = waitpid(handle, &status, blocking ? 0 : WNOHANG);
+  if (result == -1) {
+    if (auto forkServerReturn = handleForkServer()) {
+      return *forkServerReturn;
+    }
+
     CHROMIUM_LOG(ERROR) << "waitpid failed pid:" << handle
                         << " errno:" << errno;
-    if (child_exited) *child_exited = true;
-    return false;
-  } else if (result == 0) {
-    // the child hasn't exited yet.
-    if (child_exited) *child_exited = false;
+    return true;
+  }
+  if (result == 0) {
     return false;
   }
 
-  if (child_exited) *child_exited = true;
-
-  if (WIFSIGNALED(status)) {
-    switch (WTERMSIG(status)) {
-      case SIGSYS:
-      case SIGSEGV:
-      case SIGILL:
-      case SIGABRT:
-      case SIGFPE:
-        return true;
-      default:
-        return false;
-    }
+  if (WIFEXITED(status) && WEXITSTATUS(status) != 0) {
+    CHROMIUM_LOG(WARNING) << "process " << handle << " exited with status "
+                          << WEXITSTATUS(status);
+  } else if (WIFSIGNALED(status)) {
+    CHROMIUM_LOG(WARNING) << "process " << handle << " exited on signal "
+                          << WTERMSIG(status);
   }
+  return true;
 
-  if (WIFEXITED(status)) return WEXITSTATUS(status) != 0;
-
-  return false;
+#endif  // waitid
 }
 
 void FreeEnvVarsArray::operator()(char** array) {
